# backend/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plataforma Scan API Documentation",
    "description": "API documentation for the Plataforma Scan system that provides information for both mobile and web applications of the software.",
    "version": "3.5.0",
    "contact": {
      "name": "Plataforma Scan Support",
      "email": "support@plataformascan.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "https://api.plataformascan.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "OPERATOR"
            ]
          },
          "isPending": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/operators": {
      "get": {
        "summary": "List all operators",
        "description": "Retrieves a list of all operators",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of operators",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin/Manager access required"
          }
        }
      }
    },
...
    "/admin/notifications": {
      "get": {
        "summary": "List notifications",
        "description": "Retrieves a list of notifications for admins and managers",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin/Manager access required"
          }
        }
      }
    },
...
    "/admin/managers": {
      "get": {
        "summary": "List all managers",
        "description": "Retrieves a list of all managers in the system",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of managers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      },
      "post": {
        "summary": "Create manager",
        "description": "Creates a new manager account",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "verificationCode"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Manager's name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Manager's email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Manager's password"
                  },
                  "verificationCode": {
                    "type": "string",
                    "description": "Verification code for manager creation"
                  },
                  "institutionId": {
                    "type": "string",
                    "description": "ID of the institution the manager belongs to (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Manager created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Manager created successfully"
                    },
                    "manager": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or verification code"
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
    "/admin/managers/{managerId}/institution": {
      "put": {
        "summary": "Update manager's institution",
        "description": "Updates the institution a manager belongs to",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "managerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the manager to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "institutionId"
                ],
                "properties": {
                  "institutionId": {
                    "type": "string",
                    "description": "ID of the new institution"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manager's institution updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Manager or institution not found"
          }
        }
      }
    },
    "/admin/managers/{managerId}": {
      "delete": {
        "summary": "Delete manager",
        "description": "Deletes a manager by their ID",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "managerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the manager to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Manager deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Manager not found"
          }
        }
      }
    },
...
  }
}